school_review_app/requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
python-dotenv
psycopg2-binary
